refactor: move physics function to physics/mod and distinguish backend functions
